+++ test/cloud/gateway-server/src/main/java/pers/acp/spring/cloud/gateway/GateWayServerApplication.java
@@ -1,13 +1,15 @@
 package pers.acp.spring.cloud.gateway;
 
 import org.springframework.boot.SpringApplication;
-import org.springframework.cloud.client.SpringCloudApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 
 /**
  * @author zhangbin by 2018-3-10 20:45
  * @since JDK 11
  */
-@SpringCloudApplication
+@SpringBootApplication
+@EnableDiscoveryClient
 public class GateWayServerApplication {
 
     public static void main(String[] args) {