+++ doc/version_history.md
@@ -3,6 +3,8 @@
 > - [Upgrade] 升级 Spring Boot 至 2.2.2.RELEASE
 > - [Upgrade] 更新 Spring Cloud 至 Hoxton.SR1
 > - [Upgrade] 修改application.yaml中logging配置
+> - [Upgrade] 去除application.yaml中对hibernate.dialect方言的显示配置
+> - [Upgrade] MediaType.APPLICATION_JSON_UTF8_VALUE 替换为 MediaType.APPLICATION_JSON_VALUE
 > - [Upgrade] 升级依赖项
 >   - jupiter to 5.5.2
 >   - junit_platform to 1.5.2